requirements.txt
python-telegram-bot
streamlit
openai
pandas
python-dotenv
